[papi]  Refactor valdiation functions to a file
